+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Improved the list of available plugins. Now the latest stable release will be preferred.
+
 ## [2.2.0] - 2018-03-16
 ### Added
 - Automatic registering of `kernel.event_subscriber` tagged services as event subscribers.